+++ CHANGELOG.md
@@ -8,6 +8,8 @@ glance - what to expect from upgrading to a new version.
 
 ### Added
 
+- Endpoint (`/player?steam_id=`) to query player profile by Steam ID.
+
 ### Changed
 
 - Delegate various pretty-printing tasks to `silverball` instead of doing it